spv-wallet-js-client/examples/requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.11.1
bsvlib==0.10.0
certifi==2022.9.24
cffi==1.17.1
charset-normalizer==2.1.1
click==8.0.3
coincurve==20.0.0
dataclasses==0.6
Flask==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
lolpython==2.2.0
lxml==4.9.1
MarkupSafe==2.0.1
piexif==1.1.3
pillow==10.4.0
pycparser==2.22
pycryptodomex==3.22.0
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
typing-extensions==4.0.1
urllib3==1.26.12
Werkzeug==2.0.2
wikipedia==1.4.0
WTForms==3.0.1
zipp==3.6.0
